test: slim, split, and loosen backend suites toward behavior contracts
Delete implementation-locked and plugin-internal suites; loosen mock
assertions into behavior checks (KP questions, world gallery, ruleset
automation, dnd2024 director, create flow, round compression, language
support, assistant, state messages, character cards, round helpers,
access password); contract snapshot extraction now AST-based; dnd5e v2
save aliases extended to 8 migration cases.

Split oversized suites into shared harnesses plus focused files:
webapi_harness + characters/rules_worlds/create/lorebook/payments,
plugin_host_common + content_pack/security/lifecycle, and a shared
dnd2024 HTTP shim for m4/m5/m6.

Behavioral rewrites for plugin host (generation files, env isolation,
credential injection via real subprocess) and config runtime reload
(real subsystem chain); test_combat_integration unique cases merged
into test_combat_check_authority.
