+++ tests/test_contract_snapshot.py
@@ -1,11 +1,16 @@
-"""P2-N：前后端契约快照——后端关键端点返回的顶层字段必须被前端 types.ts 显式声明。
+"""前后端契约快照：后端关键端点返回的顶层字段必须被前端 types.ts 显式声明。
 
-防止"后端加字段、前端漏改"导致的静默 undefined 漂移（types.ts 的 index signature
-会让 TS 永不报错，契约测试兜住）。纯静态比对，不依赖运行环境。
+防止"后端加字段、前端漏改"导致的静默 undefined 漂移（types.ts 的 index
+signature 会让 TS 永不报错，契约测试兜住）。
+
+后端字段通过 AST 分析函数内所有 `return {...}` 字面量得到（函数重排、
+拆分变量、增加早退分支都不影响结果）；前端字段来自 types.ts 这一公开
+契约面本身。
 """
 
 from __future__ import annotations
 
+import ast
 import re
 from pathlib import Path
 
@@ -13,38 +18,26 @@ ROOT = Path(__file__).resolve().parents[1]
 
 
 def _backend_return_top_keys(path: Path, func_name: str) -> list[str]:
-    """提取后端 async 函数末尾 `return { ... }` 块里的顶层 dict key。
-
-    限定在函数体内（下一个顶层 async def 之前），并取最后一个 return 块，
-    避免早退 return 与嵌套 dict 的 key 混入。
-    """
-    text = path.read_text(encoding="utf-8")
-    start = text.find(f"async def {func_name}")
-    if start == -1:
-        raise AssertionError(f"未找到后端函数 {func_name}")
-    body_end = text.find("\nasync def ", start + 1)
-    if body_end == -1:
-        body_end = len(text)
-    ret = text.rfind("return {", start, body_end)
-    assert ret != -1, f"{func_name} 无 return dict"
-    depth = 0
-    keys: list[str] = []
-    i = ret + len("return {")
-    while i < len(text):
-        ch = text[i]
-        if ch == "{":
-            depth += 1
-        elif ch == "}":
-            depth -= 1
-            if depth == 0:
-                break
-        elif ch == '"' and depth == 0:
-            m = re.match(r'"(\w+)"\s*:', text[i:])
-            if m:
-                keys.append(m.group(1))
-                i += m.end() - 1
-        i += 1
-    return keys
+    """提取函数内所有 return 字典字面量的顶层 key（合并）。"""
+    tree = ast.parse(path.read_text(encoding="utf-8-sig"))
+    target = next(
+        (
+            node
+            for node in ast.walk(tree)
+            if isinstance(node, (ast.AsyncFunctionDef, ast.FunctionDef))
+            and node.name == func_name
+        ),
+        None,
+    )
+    assert target is not None, f"未找到后端函数 {func_name}"
+    keys: set[str] = set()
+    for node in ast.walk(target):
+        if isinstance(node, ast.Return) and isinstance(node.value, ast.Dict):
+            for key in node.value.keys:
+                if isinstance(key, ast.Constant) and isinstance(key.value, str):
+                    keys.add(key.value)
+    assert keys, f"{func_name} 未返回任何 dict 字面量"
+    return sorted(keys)
 
 
 def _ts_interface_fields(path: Path, name: str) -> set[str]:
