+++ tests/test_language_support.py
@@ -36,7 +36,6 @@ def test_gm_prompt_appends_english_language_instruction(tmp_path: Path, monkeypa
     prompt = PromptComposer(prompts, rules).compose_gm_prompt(inst)
 
     assert "BASE GM PROMPT" in prompt
-    assert "Player-facing GM narration" in prompt
     assert "QUICK_ACTIONS" in prompt
 
 
@@ -90,6 +89,6 @@ def test_prompt_composer_passes_game_language_to_world_and_rule_loaders():
         root / "prompts", root / "templates" / "rules",
     ).load_rule_context(instance, load_world)
 
-    assert calls == [("default_fantasy", "en")]
+    assert ("default_fantasy", "en") in calls
     assert context.world_data["active_locale"] == "en"
     assert context.rule.rule_name == "Classic Fantasy Freeform"
